# be/docs/swaggerSpec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PlayableCase API",
    "version": "1.0.0",
    "description": "API documentation for PlayableCase backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    },
    {
      "url": "http://13.60.247.42:5000/",
      "description": "Live server"
    }
  ],
  "components": {
    "schemas": {
      "Address": {
        "title": "Address",
        "required": [
          "owner",
          "mobileNumber",
          "houseNumber",
          "area",
          "city",
          "pincode",
          "state"
        ],
        "properties": {
          "owner": {
            "type": "schemaobjectid"
          },
          "mobileNumber": {
            "type": "number"
          },
          "houseNumber": {
            "type": "string"
          },
          "area": {
            "type": "string"
          },
          "landmark": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "state": {
            "type": "string"
          },
          "_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Product": {
        "title": "Product",
        "required": [
          "seller",
          "name",
          "description",
          "price",
          "category",
          "images"
        ],
        "properties": {
          "seller": {
            "type": "schemaobjectid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "category": {
            "type": "schemaobjectid"
          },
          "brand": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imagesPublicId": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "review": {
            "type": "array",
            "items": {
              "type": "schemaobjectid"
            }
          },
          "_id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/address/add-address": {
      "post": {
        "summary": "Add address",
        "tags": [
          "Address"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Address"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address successfully added"
          }
        }
      }
    },
    "/address/all-user-address": {
      "get": {
        "summary": "Get all addresses of the user",
        "tags": [
          "Address"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Address list"
          }
        }
      }
    },
    "/address/{addressId}": {
      "get": {
        "summary": "Get address details",
        "tags": [
          "Address"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address details"
          }
        }
      },
      "patch": {
        "summary": "Update address",
        "tags": [
          "Address"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Address"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address updated"
          }
        }
      },
      "delete": {
        "summary": "Delete address",
        "tags": [
          "Address"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address deleted"
          }
        }
      }
    },
...
    "/product/product-listing/{categoryId}": {
      "post": {
        "summary": "Add product (admin)",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "productImages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product added"
          }
        }
      }
    },
    "/product/get-product/{productId}": {
      "get": {
        "summary": "Get product details",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          }
        }
      }
    },
    "/product/update-product/{productId}": {
      "patch": {
        "summary": "Update product (admin)",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          }
        }
      }
    },
    "/product/update-product-images/{productId}": {
      "patch": {
        "summary": "Update product images (admin)",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "productImages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images updated"
          }
        }
      }
    },
    "/product/get-all-products": {
      "get": {
        "summary": "Get all products",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product list"
          }
        }
      }
    },
    "/product/change-product-category/{productId}/{categoryId}": {
      "patch": {
        "summary": "Change product category (admin)",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category changed"
          }
        }
      }
    },
    "/product/delete-product/{productId}": {
      "delete": {
        "summary": "Delete product (admin)",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          }
        }
      }
    },
...
  },
...
}